requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
youtube-transcript-api==1.2.4
biopython==1.85
entrezpy==2.1.3
nltk==3.9.1
langdetect==1.0.9
